Made colour constants, changed pattern for blank and improved structure
Made colour constants in the Bling Subsystem, so some nice colours are
easy to access.

Changed the pattern for the "Nothing's going on right now" from a
rainbow to a rainbow cycle.

Added an initialize method to the BlingPattern superclass which is run
when the pattern is first run, which is better than the previous
structure.